+++ HarmonyDB.Index/HarmonyDB.Index.Analysis.Tests/Em/TestLoop.cs
@@ -7,5 +7,5 @@ public class TestLoop : ILoop
     public required string Id { get; init; }
     public float[,] TonalityProbabilities { get; set; } = new float[Constants.TonicCount, Constants.ScaleCount];
     public (float TonicScore, float ScaleScore) Score { get; set; }
-    public required (int Tonic, Scale Scale)[] SecretTonalities { get; init; }
+    public required (byte Tonic, Scale Scale)[] SecretTonalities { get; init; }
 }
